sdk/pyproject.toml
[project]
name = "fullon_ohlcv_sdk"
version = "0.1.0"
description = "Python SDK for fullon_ohlcv_api with fullon_ohlcv object support"
authors = [{name = "Ingmar Frey", email = "ingmar.frey@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "httpx>=0.25.2",
    "websockets>=12.0",
    "fullon_ohlcv @ git+ssh://git@github.com/ingmarAvocado/fullon_ohlcv.git@main",
    "python-dotenv>=1.0.0",
]

[tool.poetry]
packages = [{include = "fullon_ohlcv_sdk"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_unimported = true
no_implicit_optional = true
show_error_codes = true